+++ src/main/java/com/launchdarkly/client/Clause.java
@@ -6,6 +6,8 @@ import com.google.gson.JsonPrimitive;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import static com.launchdarkly.client.VersionedDataKind.SEGMENTS;
+
 import java.util.List;
 
 class Clause {
@@ -16,7 +18,8 @@ class Clause {
   private List<JsonPrimitive> values; //interpreted as an OR of values
   private boolean negate;
 
-  public Clause() { }
+  public Clause() {
+  }
   
   public Clause(String attribute, Operator op, List<JsonPrimitive> values, boolean negate) {
     this.attribute = attribute;
@@ -24,8 +27,8 @@ class Clause {
     this.values = values;
     this.negate = negate;
   }
-  
-  boolean matchesUser(LDUser user) {
+
+  boolean matchesUserNoSegments(LDUser user) {
     JsonElement userValue = user.getValueForEvaluation(attribute);
     if (userValue == null) {
       return false;
@@ -51,6 +54,26 @@ class Clause {
     return false;
   }
 
+  boolean matchesUser(FeatureStore store, LDUser user) {
+    // In the case of a segment match operator, we check if the user is in any of the segments,
+    // and possibly negate
+    if (op == Operator.segmentMatch) {
+      for (JsonPrimitive j: values) {
+        if (j.isString()) {
+          Segment segment = store.get(SEGMENTS, j.getAsString());
+          if (segment != null) {
+            if (segment.matchesUser(user)) {
+              return maybeNegate(true);
+            }
+          }
+        }
+      }
+      return maybeNegate(false);
+    }
+    
+    return matchesUserNoSegments(user);
+  }
+  
   private boolean matchAny(JsonPrimitive userValue) {
     for (JsonPrimitive v : values) {
       if (op.apply(userValue, v)) {