+++ src/test/java/com/launchdarkly/client/FeatureStoreTestBase.java
@@ -1,14 +1,14 @@
 package com.launchdarkly.client;
 
+import static com.launchdarkly.client.VersionedDataKind.FEATURES;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 import java.util.HashMap;
+import java.util.Map;
 
-import org.junit.Assert;
 import org.junit.Test;
 
 public abstract class FeatureStoreTestBase<T extends FeatureStore> {
@@ -29,7 +29,9 @@ public abstract class FeatureStoreTestBase<T extends FeatureStore> {
     HashMap<String, FeatureFlag> flags = new HashMap<>();
     flags.put(feature1.getKey(), feature1);
     flags.put(feature2.getKey(), feature2);
-    store.init(flags);
+    Map<VersionedDataKind<?>, Map<String, ? extends VersionedData>> allData = new HashMap<>();
+    allData.put(FEATURES, flags);
+    store.init(allData);
   }
   
   @Test
@@ -41,14 +43,14 @@ public abstract class FeatureStoreTestBase<T extends FeatureStore> {
   @Test
   public void getExistingFeature() {
     initStore();
-    FeatureFlag result = store.get(feature1.getKey());
+    FeatureFlag result = store.get(FEATURES, feature1.getKey());
     assertEquals(feature1.getKey(), result.getKey());
   }
   
   @Test
   public void getNonexistingFeature() {
     initStore();
-    assertNull(store.get("biz"));
+    assertNull(store.get(FEATURES, "biz"));
   }
   
   @Test
@@ -57,8 +59,8 @@ public abstract class FeatureStoreTestBase<T extends FeatureStore> {
     FeatureFlag newVer = new FeatureFlagBuilder(feature1)
         .version(feature1.getVersion() + 1)
         .build();
-    store.upsert(newVer.getKey(), newVer);
-    FeatureFlag result = store.get(newVer.getKey());
+    store.upsert(FEATURES, newVer);
+    FeatureFlag result = store.get(FEATURES, newVer.getKey());
     assertEquals(newVer.getVersion(), result.getVersion());
   }
   
@@ -68,8 +70,8 @@ public abstract class FeatureStoreTestBase<T extends FeatureStore> {
     FeatureFlag oldVer = new FeatureFlagBuilder(feature1)
         .version(feature1.getVersion() - 1)
         .build();
-    store.upsert(oldVer.getKey(), oldVer);
-    FeatureFlag result = store.get(oldVer.getKey());
+    store.upsert(FEATURES, oldVer);
+    FeatureFlag result = store.get(FEATURES, oldVer.getKey());
     assertEquals(feature1.getVersion(), result.getVersion());
   }
   
@@ -79,37 +81,37 @@ public abstract class FeatureStoreTestBase<T extends FeatureStore> {
     FeatureFlag newFeature = new FeatureFlagBuilder("biz")
         .version(99)
         .build();
-    store.upsert(newFeature.getKey(), newFeature);
-    FeatureFlag result = store.get(newFeature.getKey());
+    store.upsert(FEATURES, newFeature);
+    FeatureFlag result = store.get(FEATURES, newFeature.getKey());
     assertEquals(newFeature.getKey(), result.getKey());
   }
   
   @Test
   public void deleteWithNewerVersion() {
     initStore();
-    store.delete(feature1.getKey(), feature1.getVersion() + 1);
-    assertNull(store.get(feature1.getKey()));
+    store.delete(FEATURES, feature1.getKey(), feature1.getVersion() + 1);
+    assertNull(store.get(FEATURES, feature1.getKey()));
   }
   
   @Test
   public void deleteWithOlderVersion() {
     initStore();
-    store.delete(feature1.getKey(), feature1.getVersion() - 1);
-    assertNotNull(store.get(feature1.getKey()));
+    store.delete(FEATURES, feature1.getKey(), feature1.getVersion() - 1);
+    assertNotNull(store.get(FEATURES, feature1.getKey()));
   }
   
   @Test
   public void deleteUnknownFeature() {
     initStore();
-    store.delete("biz", 11);
-    assertNull(store.get("biz"));
+    store.delete(FEATURES, "biz", 11);
+    assertNull(store.get(FEATURES, "biz"));
   }
   
   @Test
   public void upsertOlderVersionAfterDelete() {
     initStore();
-    store.delete(feature1.getKey(), feature1.getVersion() + 1);
-    store.upsert(feature1.getKey(), feature1);
-    assertNull(store.get(feature1.getKey()));
+    store.delete(FEATURES, feature1.getKey(), feature1.getVersion() + 1);
+    store.upsert(FEATURES, feature1);
+    assertNull(store.get(FEATURES, feature1.getKey()));
   }
 }