+++ CHANGELOG.md
@@ -3,6 +3,38 @@
 
 All notable changes to the LaunchDarkly Java SDK will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org).
 
+## [3.0.2] - 2018-03-01
+### Fixed
+- Improved performance when evaluating flags with custom attributes, by avoiding an unnecessary caught exception (thanks, [rbalamohan](https://github.com/launchdarkly/java-client/issues/113)).
+
+
+## [3.0.1] - 2018-02-22
+### Added
+- Support for a new LaunchDarkly feature: reusable user segments.
+
+### Changed
+- The `FeatureStore` interface has been changed to support user segment data as well as feature flags. Existing code that uses `InMemoryFeatureStore` or `RedisFeatureStore` should work as before, but custom feature store implementations will need to be updated.
+- Removed deprecated methods.
+
+
+## [3.0.0] - 2018-02-21
+
+_This release was broken and should not be used._
+
+
+## [2.6.1] - 2018-03-01
+### Fixed
+- Improved performance when evaluating flags with custom attributes, by avoiding an unnecessary caught exception (thanks, [rbalamohan](https://github.com/launchdarkly/java-client/issues/113)).
+
+
+## [2.6.0] - 2018-02-12
+## Added
+- Adds support for a future LaunchDarkly feature, coming soon: semantic version user attributes.
+
+## Changed
+- It is now possible to compute rollouts based on an integer attribute of a user, not just a string attribute.
+
+
 ## [2.5.1] - 2018-01-31
 
 ## Changed